requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
fast-agent-mcp==0.2.4
requests==2.31.0 